Move device auth to apple_tv component
