+++ homeassistant/components/services.yaml
@@ -470,3 +470,11 @@ axis:
       param:
         description: What parameter to operate on. [Required]
         example: 'package=VideoMotionDetection'
+
+apple_tv_authenticate:
+  description: Start AirPlay device authentication.
+
+  fields:
+    entity_id:
+      description: Name(s) of entities to authenticate with.
+      example: 'media_player.apple_tv'
